ancestry_list now uses 32-bit indexing.

+++ fwdpp/ts/simplification/simplification.hpp
@@ -193,7 +193,7 @@ namespace fwdpp
                 }
             };
 
-            using ancestry_list = nested_forward_lists<segment, std::int64_t, -1>;
+            using ancestry_list = nested_forward_lists<segment, std::int32_t, -1>;
 
             template <typename TableCollectionType> struct simplifier_internal_state
             /// Holds data needed during tree sequence simplification
